Move individual panels to 'panel.py'
Allows having animations features for indivudal things without circ refs.
